Clean up sparse solver module
