+++ gEconpy/solvers/sparse_root/__init__.py
@@ -1,10 +1,7 @@
-from gEconpy.solvers.sparse_root.chord import Chord
 from gEconpy.solvers.sparse_root.dogleg import SparseDogleg
 from gEconpy.solvers.sparse_root.gauss_newton import GaussNewtonTrustRegion
-from gEconpy.solvers.sparse_root.krylov import InexactNewtonKrylov
 from gEconpy.solvers.sparse_root.levenberg_marquardt import LevenbergMarquardt
-from gEconpy.solvers.sparse_root.newton import NewtonArmijo
-from gEconpy.solvers.sparse_root.nonmonotone import NewtonNonmonotone
+from gEconpy.solvers.sparse_root.line_search import Chord, InexactNewtonKrylov, NewtonArmijo, NewtonNonmonotone
 from gEconpy.solvers.sparse_root.sparse_root import sparse_root
 
 __all__ = [
